requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
reportlab==4.0.4
reportlab-cjk==0.1.0
PyMySQL==1.1.0
python-dotenv==1.0.0